small DAGStoreWrapper code reorg.
